+++ R/maven.R
@@ -1,15 +1,18 @@
 ## Java Virtal Machine functions ----
 
 # find java home for JDK and sets system variable
-.java_home = function(set = TRUE, quiet = getOption("rmaven.quiet",FALSE), require_jdk = FALSE) {
+.java_home = function(set = TRUE, quiet = getOption("rmaven.quiet",TRUE), require_jdk = FALSE) {
   jh = getOption("rmaven.java_home",NA)
   if (is.na(jh)) jh = Sys.getenv("JAVA_HOME", unset=NA)
   if (is.na(jh)) {
-    try({start_jvm(quiet = TRUE)},silent=TRUE)
+    if (!rJava::.jniInitialized) {
+      warning("Initialising rJava with default parameters to check for JAVA_HOME. This probably should have been called after rJava::.jinit() or rmaven::start_jvm()")
+      try({start_jvm(quiet = TRUE)},silent=TRUE)
+    }
     jh = tryCatch({rJava::.jcall( 'java/lang/System', 'S', 'getProperty', 'java.home' )},error = function(e) NA)
   }
   if(is.na(jh)) jh = utils::tail(unlist(stringr::str_split(Sys.getenv("LD_LIBRARY_PATH"),":")),1)
-  if(is.na(jh)) stop("Could not determine JAVA_HOME from rJava, LD_LIBRARY_PATH, or Sys.getenv")
+  if(is.na(jh)) stop("Could not determine JAVA_HOME from getOption('rmaven.java_home'), Sys.getenv('JAVA_HOME'), Sys.getenv('LD_LIBRARY_PATH'), or via 'rJava'")
 
   jh_orig = jh
   jh_parent = fs::path_dir(jh)
@@ -30,7 +33,7 @@
       jh = jh_parent
       jh_parent = fs::path_dir(jh)
     }
-    stop("Couldn't find 'bin/javac(.exe)' or 'bin/java(.exe)' in any parent directories starting at ",jh_orig,", please set options('rmaven.java_home'=...) to the root of a JDK (the directory above 'bin/javac').")
+    stop("Couldn't find 'bin/javac(.exe)' or 'bin/java(.exe)' in any parent directories starting at ",jh_orig,", please set options('rmaven.java_home'=...) to the root of a JDK (the directory containing 'bin/javac').")
   }
   if (set) Sys.setenv("JAVA_HOME"=jh)
   return(jh)
@@ -58,10 +61,14 @@
 
 #' Start an `rJava` `JVM` with or without debugging options
 #'
-#' This does not do anything if the `JVM` has already been started.
+#' This does not do anything if the `JVM` has already been started. Otherwise starts the JVM via `rJava` with a set of options
+#' Additional JVM options (beyond debugging) can be set with the `options("java.parameters"=c("-Xprof","-Xrunhprof"))`
 #'
 #' @param debug turn on debugging
-#' @param quiet don't report messages (defaults to `getOption("rmaven.quiet")` or FALSE)
+#' @param quiet don't report messages (defaults to `getOption("rmaven.quiet")` or TRUE)
+#' @param max_heap optional. if a string like `"2048m"` the `-Xmx` option value to start the JVM - if a string like `"75%"` the `-XX:MaxRAMPercentage`, if a numeric - number of megabytes.
+#' @param thread_stack optional. sensible values range from "1m" to "128m" (max is "1Gb"). Can be important with serialising deeply nested structures.
+#' @param ... any other named parameters are passes as `-name` or `-name=value` if value is a character
 #'
 #' @return nothing - called for side effects
 #' @export
@@ -72,28 +79,65 @@
 #' # this may try to rebind debugging port
 #' start_jvm(debug = TRUE)
 #' }
-start_jvm = function(debug = FALSE, quiet = getOption("rmaven.quiet",FALSE)) {
+start_jvm = function(debug = FALSE, quiet = getOption("rmaven.quiet",TRUE), max_heap = NULL, thread_stack = NULL, ...) {
+  opts = getOption("java.parameters")
+  if (!is.null(max_heap)) {
+    # get rid of startup Xmx option
+    opts = opts[substr(opts,1,4)!="-Xmx"]
+    if ("%" == substr(max_heap,start = nchar(max_heap),nchar(max_heap))) {
+      # ends with %
+      max_heap = substr(max_heap,start = 1,nchar(max_heap)-1)
+      opts = c(opts,sprintf("-XX:MaxRAMPercentage=%s.0",max_heap))
+    } else if (is.numeric(max_heap)) {
+      unit = "m" # if (max_heap > 32) "m" else "g"
+      # assume number of megabytes
+      opts = c(opts,sprintf("-Xmx%1.0f%s",max_heap,unit))
+    } else {
+      opts = c(opts,sprintf("-Xmx%s",max_heap))
+    }
+  }
+  if (!is.null(thread_stack)) opts = c(opts,sprintf("-Xss%s",thread_stack))
+
+  named = rlang::dots_list(..., .homonyms = "error")
+  if (length(named) > 0) {
+    named = named[!unlist(lapply(named, is.null))]
+    named = named[names(named) != ""]
+    opts2 = ifelse(
+      unlist(named)=="",
+      paste0("-",names(named)), # no option value needed
+      paste0("-",names(named),"=",unlist(named))
+    )
+  } else {
+    opts2 = NULL
+  }
+
+  opts = unname(c(opts, opts2))
   tryCatch({
     if (!rJava::.jniInitialized) {
       if (debug) {
       # pass in debug options
-        rJava::.jinit(parameters=c(getOption("java.parameters"),"-Xdebug","-Xrunjdwp:transport=dt_socket,address=8998,server=y,suspend=n"), silent = TRUE, force.init = TRUE)
-        if(!quiet) message("java debugging initialised on port 8998")
+        rJava::.jinit(parameters=c(opts,"-Xdebug","-Xrunjdwp:transport=dt_socket,address=8998,server=y,suspend=n"), silent = TRUE, force.init = TRUE)
+        if(!quiet) message("java debugging initialised on port 8998 with options: ",opts)
       } else {
-        rJava::.jinit(parameters=getOption("java.parameters"),silent = TRUE, force.init = FALSE)
+        rJava::.jinit(parameters=opts,silent = TRUE, force.init = FALSE)
       }
+    } else {
+      if(!quiet) warning("The JVM was already initialised when start_jvm(...) was called. The following parameters were not applied: ", opts)
     }
-  }, error = function(e) stop("Java cannot be initialised: ",e$message))
+  }, error = function(e) {
+    warning("")
+    stop("Java cannot be initialised: ",e$message)
+  })
 }
 
 ## File functions ----
 
-#' Find location of all the jars in a particular package
+#' Find location of all the jars in a particular package.
 #'
 #' @param package_name the R package name
 #' @param types the jar types to look for in the package: one of `all`,`thin-jar`,`fat-jar`,`src`
 #'
-#' @return a vector of package jars
+#' @return a vector of paths to jar files in the package
 #' @export
 #'
 #' @examples
@@ -189,9 +233,9 @@ print.coordinates = function(x,...) {
 #' @param groupId the maven `groupId`
 #' @param artifactId the maven `artifactId`
 #' @param version the maven version
-#' @param ... other parameters ignored apart from packaging (`jar`,`war`,`pom` or `ejb`) and classifier (`tests`, `client`, `sources`, `javadoc`, `jar-with-dependencies`, `src`)
+#' @param ... other parameters ignored apart from `packaging` (one of `jar`,`war`,`pom` or `ejb`) and `classifier` (one of `tests`, `client`, `sources`, `javadoc`, `jar-with-dependencies`, `src`)
 #'
-#' @return a coordinates object containing the coordinates
+#' @return a coordinates object containing the Maven artifact coordinates
 #' @export
 #'
 #' @examples
@@ -338,12 +382,12 @@ as.coordinates = function(groupId, artifactId, version, ...) {
 
 #' Sets the local maven repository location
 #'
-#' This writes a maven repository location to a temporary 'settings.xml' file which persists only for the R session.
-#' The location of the maven repository is either specified here, or can be defined by the 'options("rmaven.m2.repository"=...)' option.
-#' If neither of these is provided, the location will revert to a default location within the 'rmaven' cache. (Approved by CRAN for a local cache location)
-#' e.g. on 'Linux' this will default to '~/.cache/rmaven/.m2/repository/'
+#' This writes a maven repository location to a temporary `settings.xml` file which persists only for the R session.
+#' The location of the maven repository is either specified here, or can be defined by the `options("rmaven.m2.repository"=...)` option.
+#' If neither of these is provided, the location will revert to a default location within the `rmaven` cache. (Approved by CRAN for a local cache location)
+#' e.g. on 'Linux' this will default to `~/.cache/rmaven/.m2/repository/`
 #'
-#' @param repository_location a file path (which will be expanded to a full path) where the repository should be based, e.g. '~/.m2/repository/'. Defaults to a sub-directory of the 'rmaven' cache.
+#' @param repository_location a file path (which will be expanded to a full path) where the repository should be based, e.g. `~/.m2/repository/`. Defaults to a sub-directory of the `rmaven` cache.
 #' @param settings_path the file path of the settings.xml to update (generally the supplied default is what you want to use)
 #'
 #' @return the new repository location (expanded)
@@ -381,14 +425,14 @@ set_repository_location = function(
 #' Get the location of the Maven repository
 #'
 #' In general this function is mainly for internal use but maybe handy for debugging.
-#' The maven repository location can be defined by 'set_repository_location(...)' or through the option
-#' 'options("rmaven.m2.repository"=...)' option but defaults to a '.m2/repository' directory in the 'rmaven' cache directory.
+#' The maven repository location can be defined by `set_repository_location(...)` or through the option
+#' `options("rmaven.m2.repository"=...)` option but defaults to a `.m2/repository` directory in the `rmaven` cache directory.
 #' This is not the default location for Maven when used from Java writing to the default Maven directory in user space is
-#' forbidden by CRAN policies. The result of this is that 'rmaven' will have to unnecessarily download additional copies of java
+#' forbidden by CRAN policies. The result of this is that `rmaven` will have to unnecessarily download additional copies of java
 #' libraries, onto the users computer and cannot re-use already cached copies. This is more of an issue for developers rather
 #' than users.
 #'
-#' @param settings_path the file path of the settings.xml to update (generally the supplied default is what you want to use)
+#' @param settings_path the file path of the `settings.xml` to update (generally the supplied default is what you want to use)
 #' @export
 #'
 #' @examples
@@ -413,11 +457,13 @@ get_repository_location = function(settings_path = .settings_path()) {
 }
 
 
-#' Clear out the 'rmaven' cache
+#' Clear out the `rmaven` cache
 #'
-#' Deletes all content in the 'rmaven' cache. This should not be necessary, but never
+#' Deletes all content in the `rmaven` cache. This should not be necessary, but never
 #' say never, and if there is really a problem with the cache, then deleting it may be the
-#' best thing.
+#' best thing. This will wait for confirmation from the user. If running unattended the
+#' `options("rmaven.allow.cache.delete"=TRUE)` must be set for the action to occur, otherwise
+#' it will generate a warning and do nothing.
 #'
 #' @return nothing, called for side effects
 #' @export
@@ -435,7 +481,7 @@ clear_rmaven_cache = function() {
     if (getOption("rmaven.allow.cache.delete",FALSE)) {
       fs::dir_delete(dir)
     } else {
-      message("the option rmaven.allow.cache.delete was not TRUE (or not set)")
+      warning("the option rmaven.allow.cache.delete was not TRUE (or not set)")
       message("so we can't delete the rmaven cache in an unattended script")
     }
   } else {
@@ -456,7 +502,7 @@ clear_rmaven_cache = function() {
 
 # internal function
 # loads a maven wrapper distribution from the internet and unzips it into the rmaven working directory
-.load_maven_wrapper = function(quiet = getOption("rmaven.quiet",FALSE)) {
+.load_maven_wrapper = function(quiet = getOption("rmaven.quiet",TRUE)) {
   dir = .working_dir()
   if (!file.exists(.working_file("mvnw"))) {
 
@@ -488,13 +534,15 @@ clear_rmaven_cache = function() {
 }
 
 # verbosity settings
+# defaults to quiet
 .quietly = function(verbose = "normal") {
   if (verbose == "quiet") return(TRUE)
   if (verbose == "debug") return(FALSE)
-  return(getOption("rmaven.quiet",FALSE))
+  return(getOption("rmaven.quiet",TRUE))
 }
 
-# verbosity settings
+# debug settings
+# defaults to
 .debug = function(verbose = "normal") {
   if (verbose == "debug") return(TRUE)
   if (verbose == "quiet") return(FALSE)
@@ -503,18 +551,18 @@ clear_rmaven_cache = function() {
 
 #' Executes a maven goal
 #'
-#' Maven goals may be executed with or without a pom.xml file. Some maven goals (e.g. compilation)
-#' require the use of a `JDK`.
+#' Maven goals are defined either as lifecycle goals (e.g. "clean", "compile") or as plugin goals (e.g. "help:system"). Some Maven goals may be executed without a `pom.xml` file, others require one.
+#' Some maven goals (e.g. compilation) require the use of a `JDK`.
 #'
 #' @param goal the goal of the `mvn` command ( can be multiple ) e.g. `c("clean","compile")`
 #' @param opts provided options in the form `c("-Doption1=value2","-Doption2=value2")`
-#' @param pom_path optional. the path to a pom.xml file for goals that operate on one
-#' @param quiet should output from maven be suppressed? (-q flag)
-#' @param debug should output from maven be verbose? (-X flag)
+#' @param pom_path optional. the path to a `pom.xml` file for goals that need one.
+#' @param quiet should output from maven be suppressed? (`-q` flag)
+#' @param debug should output from maven be verbose? (`-X` flag)
 #' @param verbose how much output from maven, one of "normal", "quiet", "debug"
-#' @param require_jdk does the goal you are executing require a `JDK` (e.g. compilation)
-#' @param settings the path to a settings.xml file controlling Maven. The default is a configuration with a local repository in the 'rmaven' cache directory (and not the Java maven repository).
-#' @param ... named parameters are passed to maven as options in the form `-Dname=value`
+#' @param require_jdk does the goal you are executing require a `JDK` (e.g. compilation does, fetching artifacts and calculating classpaths does not)
+#' @param settings the path to a `settings.xml` file controlling Maven. The default is a configuration with a local repository in the `rmaven` cache directory (and not the Java maven repository).
+#' @param ... non-empty named parameters are passed to maven as options in the form `-Dname=value`
 #'
 #' @return nothing, invisibly
 #' @export
@@ -522,7 +570,7 @@ clear_rmaven_cache = function() {
 #' @examples
 #' \donttest{
 #' # This code can take quite a while to run as has to
-#' # download a lot of plugins, especially on first run
+#' # download a lot of plugins, especially on first run on a clean system
 #' execute_maven("help:system")
 #' }
 execute_maven = function(goal, opts = c(), pom_path=NULL, quiet=.quietly(verbose), debug=.debug(verbose), verbose = c("normal","debug","quiet"), require_jdk=FALSE, settings = .settings_path(), ...) {
@@ -548,21 +596,12 @@ execute_maven = function(goal, opts = c(), pom_path=NULL, quiet=.quietly(verbose
   ) #, paste0("-f \"",pomPath,"\""))
   if (quiet) args = c(args, "-q")
   if (debug) args = c(args, "-X")
+  # make sure JAVA_HOME is set
   .java_home(quiet=TRUE)
+
   # changing the wd is required due to an issue in Mvnw.cmd on windows.
+  # N.b. this is probably not true as may just have been a quotes issue
   wd = getwd()
-
-    # if (Platform.os == "windows") {
-    # wrapper_props = fs::path(fs::path_dir(pom_path),".mvn/wrapper/maven-wrapper.properties")
-    # # Windows mvnw.cmd looks for this in the directory the pom is in.
-    # if (!file.exists(wrapper_props)) {
-    #   write(c(
-    #     "distributionUrl=https://repo.maven.apache.org/maven2/org/apache/maven/apache-maven/3.3.9/apache-maven-3.3.9-bin.zip",
-    #     "wrapperUrl=https://repo.maven.apache.org/maven2/org/apache/maven/wrapper/maven-wrapper/3.1.1/maven-wrapper-3.1.1.jar"
-    #   ), wrapper_props)
-    # }
-    # }
-
   # change the working directory
   if(!is.null(pom_path)) {
     setwd(fs::path_dir(pom_path))
@@ -578,22 +617,21 @@ execute_maven = function(goal, opts = c(), pom_path=NULL, quiet=.quietly(verbose
   invisible(NULL)
 }
 
-#' Fetch an artifact from a repository into the local .m2 cache
+#' Fetch an artifact from a remote repository into the local .m2 cache
 #'
-
 #' @param groupId optional, the maven `groupId`,
 #' @param artifactId optional, the maven `artifactId`,
 #' @param version optional, the maven version,
 #' @param ... other maven coordinates such as classifier or packaging
 #' @param coordinates optional, coordinates as a coordinates object,
 #' @param artifact optional, coordinates as an artifact string `groupId:artifactId:version[:packaging[:classifier]]` string
-#' @param repoUrl the URLs of the repositories to check (defaults to maven central)
+#' @param repoUrl the URLs of the repositories to check (defaults to maven central, sonatype snapshots and jitpack, defined in options("rmaven.default_repos"))
 #' @param coordinates optional, but if not supplied `groupId` and `artifactId` must be, coordinates as a coordinates object (see as.coordinates())
 #' @param artifact optional, coordinates as an artifact string `groupId:artifactId:version[:packaging[:classifier]]` string
 #' @param nocache normally artifacts are only fetched if required, `nocache` forces fetching
 #' @param verbose how much output from maven, one of "normal", "quiet", "debug"
 #'
-#' @return the .m2 path of the artifact
+#' @return the path of the artifact within the local maven cache
 #' @export
 #'
 #' @examples
@@ -610,7 +648,7 @@ fetch_artifact = function(
     ...,
     coordinates = NULL,
     artifact = NULL,
-    repoUrl = getOption("rmaven.default_repos"),
+    repoUrl = .default_repos(),
     nocache = FALSE,
     verbose = c("normal","quiet","debug")
 ) {
@@ -648,14 +686,16 @@ fetch_artifact = function(
 
 #' Copy an artifact from a repository to a local directory
 #'
+#' This essentially runs a `maven-dependency-plugin:copy` goal to copy a JAR file from a remote repository to a local directory.
+#'
 #' @param groupId optional, the maven `groupId`,
 #' @param artifactId optional, the maven `artifactId`,
 #' @param version optional, the maven version,
 #' @param ... other maven coordinates such as classifier or packaging
 #' @param coordinates optional, coordinates as a coordinates object,
 #' @param artifact optional, coordinates as an artifact string `groupId:artifactId:version[:packaging[:classifier]]` string
-#' @param repoUrl the URLs of the repositories to check (defaults to maven central & `Sonatype snaphots`)
-#' @param outputDirectory optional path, defaults to the rmaven cache directory
+#' @param repoUrl the URLs of the repositories to check (defaults to maven central, `Sonatype snaphots` and `jitpack`)
+#' @param outputDirectory optional path, defaults to the `rmaven` cache directory
 #' @param nocache normally artifacts are only fetched if required, `nocache` forces fetching
 #' @param verbose how much output from maven, one of "normal", "quiet", "debug"
 #'
@@ -677,7 +717,7 @@ copy_artifact = function(
     coordinates = NULL,
     artifact = NULL,
     outputDirectory = .working_dir(artifact),
-    repoUrl = getOption("rmaven.default_repos"),
+    repoUrl = .default_repos(),
     nocache = FALSE,
     verbose = c("normal","quiet","debug")
 ) {
@@ -781,7 +821,8 @@ copy_artifact = function(
 #' Resolve the `classpath` for an artifact
 #'
 #' This calculates the dependencies for an artifact which may be specified either as a set of maven coordinates (in which case the
-#' artifact is downloaded) or as a path to a jar file containing a pom.xml (e.g. a compiled jar file, a compiled jar-with-dependencies, or a assembled `...-src.jar`)
+#' artifact is downloaded, and included in the `classpath`) or as a path to a jar file containing a pom.xml (e.g. a compiled jar file,
+#' a compiled jar-with-dependencies, or a assembled `...-src.jar`)
 #' The resulting file paths which will be in the maven local cache are checked on the file system.
 #'
 #' @param groupId the maven `groupId`, optional
@@ -790,7 +831,7 @@ copy_artifact = function(
 #' @param ... passed on to as.coordinates()
 #' @param coordinates the maven coordinates, optional (either `groupId`,`artifactId` and 'version' must be specified, or 'coordinates', or 'artifact')
 #' @param artifact optional, coordinates as an artifact string `groupId:artifactId:version[:packaging[:classifier]]` string
-#' @param path the path to the source directory, pom file or jar file. if blank the
+#' @param path the path to the source directory, pom file or jar file. if not given `rmaven` will get the artifact from the maven central repositories
 #' @param include_self do you want include this path in the `classpath`. optional, if missing the path will be included if it is a regular jar, or a fat jar, otherwise not.
 #' @param nocache do not used cached version, by default we use a cached version of the `classpath` unless the `pom.xml` is newer that the cached `classpath`.
 #' @param verbose how much output from maven, one of "normal", "quiet", "debug"
@@ -803,13 +844,17 @@ copy_artifact = function(
 #' # This code can take quite a while to run as has
 #' # to download a lot of plugins, especially on first run
 #'
+#' # classpath would be cached if possible
 #' resolve_dependencies(groupId = "commons-io", artifactId = "commons-io", version="2.11.0")
 #'
+#' # forcing download and classpath calculation of an artifact
 #' resolve_dependencies(artifact = "org.junit.jupiter:junit-jupiter-api:5.9.0", nocache=TRUE)
 #'
+#' # find the test jar in this package and calculate its stated dependencies
 #' resolve_dependencies(path=
 #'   system.file("testdata/test-project-0.0.1-SNAPSHOT.jar",package="rmaven"))
 #'
+#' # find the test source code jar in this package and calculate its stated dependencies
 #' resolve_dependencies(path=
 #'   system.file("testdata/test-project-0.0.1-SNAPSHOT-src.jar",package="rmaven"))
 #' }
@@ -924,7 +969,7 @@ resolve_dependencies = function(
 
   verbose = match.arg(verbose)
 
-  if (!fs::file_exists(path)) stop("we didn't something to compile at ",path)
+  if (!fs::file_exists(path)) stop("The was nothing to compile at: ",path)
 
   if (fs::path_ext(path) == "xml") {
     pom_path = path
@@ -943,7 +988,7 @@ resolve_dependencies = function(
 
   if (nocache) unlink( fs::path(project_dir,"target"), recursive = TRUE)
 
-  if (pom_path %>% .is_newer_than(target_jar)) {
+  if (path %>% .is_newer_than(target_jar)) {
     if (.quietly(verbose)) message("Compiling Java library, please be patient.")
     execute_maven(
       pom_path,
@@ -953,6 +998,8 @@ resolve_dependencies = function(
       require_jdk = TRUE,
       ...
     )
+  } else {
+    if (!.quietly(verbose)) message("Compilation aready performed")
   }
 
   if(!file.exists(target_jar)) stop("could not compile java and assemble file: ",fs::path_file(target_jar))
@@ -966,13 +1013,14 @@ resolve_dependencies = function(
 #' Compilation will package the Java source code in to a Jar file for further use. It will resolve dependencies and
 #' optionally package them into a single `uber jar` (using maven assembly).
 #'
-#' @param path the path to - a source code directory containing a pom.xml file, a `...-src.jar` assembled by the maven assembly plugin, or a `pom.xml` file
+#' @param path the path to - either a java source code directory containing a `pom.xml` file, the `pom.xml` file itself, or a `...-src.jar` assembled by the maven assembly plugin,
 #' @param nocache normally compilation is only performed if the input has changed. `nocache` forces recompilation
 #' @param verbose how much output from maven, one of "normal", "quiet", "debug"
-#' @param with_dependencies compile the Java code to a '...-jar-with-dependencies.jar' including transitive dependencies which is easier to embed
-#' @param ... passed to execute_maven(...), e.g. could include settings parameter
+#' @param with_dependencies compile the Java code to a '...-jar-with-dependencies.jar' including transitive dependencies which may be easier to embed into R code
+#' as does not need a classpath (however may be large if there are a lot of dependencies)
+#' @param ... passed to `execute_maven(...)`, e.g. could include `settings` parameter
 #'
-#' @return the path to the compiled 'jar' file.
+#' @return the path to the compiled 'jar' file. If this is a fat jar this can be passed straight to `rJava`, otherwise an additional `resolve_dependencies(...)` call is required
 #' @export
 #'
 #' @examples
@@ -1015,9 +1063,11 @@ compile_jar = function(path, nocache = FALSE, verbose = c("normal", "quiet", "de
   return(target_jar)
 }
 
-
-options("rmaven.default_repos" = c(
-  "https://repo1.maven.org/maven2/",
-  "https://s01.oss.sonatype.org/content/repositories/snapshots/",
-  "https://jitpack.io"
-))
+# TODO: put this into the settings.xml
+.default_repos = function() {
+  getOption("rmaven.default_repos", default = c(
+    "https://repo1.maven.org/maven2/",
+    "https://s01.oss.sonatype.org/content/repositories/snapshots/",
+    "https://jitpack.io"
+  ))
+}